+++ src/Modules/HRManagement.Modules.Staff/Features/Services/RoleCreateOrUpdateDto.cs
@@ -6,12 +6,12 @@ namespace HRManagement.Modules.Staff.Features.Services;
 
 public class RoleCreateOrUpdateDto
 {
-    public byte Id { get; init; }
+    public int Id { get; init; }
 
     public RoleName Name { get; init; }
 
     public Role RoleToUpdate { get; set; }
 
-    public byte? ManagerRoleId { get; init; }
+    public int? ManagerRoleId { get; init; }
     public Maybe<Role> ManagerRoleOrNothing { get; set; } = Maybe<Role>.None;
 }
